backend/requirements.txt
fastapi
uvicorn
pydantic
openai
google-api-python-client
crawl4ai
pyyaml
python-dotenv
jinja2
markdown
beautifulsoup4
slowapi
redis
pydantic_settings